Improves some optimizations
Loop-invariant-code-motion now also hoists (some) non-constant loop
invariant code out of loops into the dominator (pre-header) node.

This hoists e.g. 12 instructions out of a loop for every code with the following pattern:

    for(...) {
        barrier(...);
        // [...] some more code
    }

Also improves the redundant code eliminator by removing ANDing zero-extended
bytes by 0xFF (as e.g. happens for calculation of local IDs and sizes).

+++ src/tools/Emulator.cpp
@@ -2208,7 +2208,7 @@ EmulationResult tools::emulate(const EmulationData& data)
     auto it = instructions.begin() +
         static_cast<std::vector<qpu_asm::Instruction>::difference_type>(
             (kernelInfo->getOffset() - module.kernelInfos.front().getOffset()).getValue());
-    result.instrumentation.reserve(kernelInfo->getLength().getValue());
+    result.instrumentation.reserve(static_cast<std::size_t>(kernelInfo->getLength().getValue()));
     while(true)
     {
         result.instrumentation.emplace_back(instrumentation[&(*it)]);
